Refactoring: Extract ExperimientStore
Extracts the responsibility of finding experiments to another class
