+++ lib/split/experiment.rb
@@ -38,31 +38,20 @@ module Split
     end
 
     def self.all
-      Split.redis.smembers(:experiments).map {|e| find(e)}
+      ExperimentCatalog.all
     end
 
     # Return experiments without a winner (considered "active") first
     def self.all_active_first
-      all.partition{|e| not e.winner}.map{|es| es.sort_by(&:name)}.flatten
+      ExperimentCatalog.all_active_first
     end
 
     def self.find(name)
-      if Split.redis.exists(name)
-        obj = self.new name
-        obj.load_from_redis
-      else
-        obj = nil
-      end
-      obj
+      ExperimentCatalog.find(name)
     end
 
     def self.find_or_create(label, *alternatives)
-      experiment_name_with_version, goals = normalize_experiment(label)
-      name = experiment_name_with_version.to_s.split(':')[0]
-
-      exp = self.new name, :alternatives => alternatives, :goals => goals
-      exp.save
-      exp
+      ExperimentCatalog.find_or_create(label, *alternatives)
     end
 
     def save
@@ -246,17 +235,6 @@ module Split
     end
 
     protected
-
-    def self.normalize_experiment(label)
-      if Hash === label
-        experiment_name = label.keys.first
-        goals = label.values.first
-      else
-        experiment_name = label
-        goals = []
-      end
-      return experiment_name, goals
-    end
 
     def experiment_config_key
       "experiment_configurations/#{@name}"
